requirements.txt
discord==1.0.1
discord.py==1.7.1
PyNaCl==1.4.0
python-dotenv==0.17.0
lavalink